fix(infrastructure): Getting scraping to work with Fly.io (#64)
This commit moves the scraping mechanism to a Github workflow instead of using an internal mechanism (like the Heroku scheduler plugin). I am using the GitHub Actions scheduling API to schedule the scraping for twice a day. I use a separate GitHub Action in the workflow to actually run the CURL command.

* Adding scrape Github Actions workflow

* Updating Comments

* Adding workflow trigger for branch

* Fixing branch name in scrape workflow

* Fixing scrape workflow cron syntax

* Fixing spacing in Scrape yaml file

* Removing Temp Scrape on Push