[Issue 69] Fixing Migration To Oracle Cloud (#71)
These changes will move the building of the docker image to GitHub Actions instead of on the Oracle Cloud VM. The original Dokku plugin was pushing the repository to the VM and then using docker on the VM to build the image and then run it. This was failing from OOM issues and the image was never built. Moving the building to GitHub Actions moves the build process to GitHub machines which have way more resources to build the application. It also pushes the images to the GitHub container registry. The Dokku deploy step just tells Dokku to pull the new image from the container registry. These changes include:
- new workflow file for deploy
  - new job for building the image in GitHub Actions, tagging the image, and pushing it to the container registry
  - moved job for deploying to VM with Dokku, but tweaked so it only pulls build image (not whole repo)
- removed Fly.toml file